requirements.txt
docx_mailmerge==0.5.0
comtypes==1.1.7
pandas==0.24.2